+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - More documentation in the README.
 - More tests.
+- Check that `get_context_data(self, **kwargs)` is not being overridden.
 
 ## [0.2.0] - 2017-08-07
 ### Added